% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div_quick.R
\name{beta_div_quick}
\alias{beta_div_quick}
\title{Calculate Beta Diversity Quickly}
\usage{
beta_div_quick(Y, method = c("hellinger", "jaccard", "sorensen", "ochiai"))
}
\arguments{
\item{Y}{a matrix of presence-absence data; species are columns, sites are rows}

\item{method}{the method for calculating beta diversity}
}
\value{
Returns a single numeric value that is the total beta diversity of the matrix (SS_total/(n-1))
}
\description{
Calculates beta diversity using a modified version of Legendre & Caceres's \code{beta.div}
}
\details{
in Ecology Letters (2013) 16: 951-963 "Beta diversity as the variance of community data: dissimilarity coefficients and partitioning" a function \code{beta.div} is provided in the supplement. This function is an abbreviated version that can be used for 4 out of the original's 20 options, and only returns the total beta diversity (original returned many other values). This function was written to be much faster than the original, as the original can be quite slow if being run for 10's of thousands of matrices (think Bayesian posterior of MSOM results).
}
\examples{
Y <- matrix(rbinom(18, 1, 0.5), ncol=3)
bd1 <- beta_div_quick(Y, "hellinger")
bd2 <- beta_div_quick(Y, "jaccard")

}
\seealso{
\link{process_msomStatic}
}

