% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{comm_master}
\alias{comm_master}
\title{Master Community-Level Data}
\format{data.table}
\usage{
comm_master
}
\description{
Master data set for community level metrics related to beta diversity, species richness, environmental values, and cross-species averages of MSOM parameters
}
\details{
Many of the columns in this data set are cross-species averages of columns by the same (or similar) name in \code{\link{spp_master}}.

\tabular{lll}{
\code{reg} \tab character \tab region name \cr
\code{year} \tab integer \tab year \cr
\code{method} \tab character \tab method for standardization for beta diversity calculation (see \code{\link{beta_div_quick}}) \cr
\code{beta_div_mu} \tab numeric \tab for each posterior iteration from MSOM Z, beta diversity is calculated; this is the average of those posterior calculations \cr
\code{beta_div_sd} \tab numeric \tab standard deviation of posterior beta diversity \cr
\code{Omega} \tab numeric \tab posterior average of the MSOM Omega parameter \cr
\code{reg_rich} \tab numeric \tab posterior average of MSOM richness estimate \cr
\code{n_col} \tab integer \tab number of observed colonizers this year (see \code{\link{get_colonizers}}) \cr
\code{n_ext} \tab integer \tab number of species that were observed this year, but will not be observed next year (see \code{\link{get_colonizers}}) \cr
\code{n_pers} \tab integer \tab number of species present this year that are also present in adjacent sampling years (i.e., = richness - (n_col+n_ext)) (see \code{\link{get_colonizers}}) \cr
\code{bt_ann} \tab numeric \tab cross-stratum average bottom temperature for current year (same as in \code{\link{spp_master}}) \cr
\code{naive_rich} \tab numeric \tab observed species richness \cr
\code{bt_opt_avg, bt_tol_avg, detect_mu_avg, detect_mu} \tab numeric \tab cross-species average of column by same name in \code{\link{spp_master}} \cr
\code{propStrata_avg} \tab numeric \tab cross species average of \code{propStrata} in \code{\link{spp_master}} \cr
}
}
\seealso{
\code{\link{spp_master}}, \code{\link{mapDat}}
}
\keyword{datasets}

