% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataArray.R
\name{dataArray}
\alias{dataArray}
\title{Array Format for MSOM Data}
\usage{
dataArray(X, bioName = "abund", covNames = c("btemp", "nAgg2"))
}
\arguments{
\item{X}{A data.table of trawl data}

\item{bioName}{A character vector of length one specifying a column in \code{X} to be used as the biological response variable}

\item{covNames}{Names of columns in \code{X} to be used as covariates}
}
\value{
A list of length \code{length(c(bioName, covNames))} whose elements are named accordingly, and the contents of each element contain the corresponding arrays. Array dimensions for bioName are \code{stratum~K~spp~year}. Array dimensions for covNames are \code{stratum~K~year}. Grand dimension names for the bioName array are defaults for \code{trawlData::trawlCast}; names for covNames arrays are as in formula.
}
\description{
Given a data.table of trawl data , turn biological response and covariates into acceptable array format for MSOM analysis
}

