% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ebs.a}
\alias{ebs.a}
\title{ebs.a}
\format{A dim = 184 x 21 data.table data.frame:
\tabular{rlll}{
[,1] \tab aggStrat \tab character \tab stratum used in spatial aggregation\cr
[,2] \tab time_lvl \tab character \tab the temporal factor\cr
[,3] \tab spp \tab character \tab taxonomic identifier (species name)\cr
[,4] \tab wtcpue \tab numeric \tab weight per unit effort\cr
[,5] \tab abund \tab integer \tab number of aggrated hauls with wtcpue >= 0\cr
[,6] \tab stemp \tab numeric \tab surface tempeature\cr
[,7] \tab btemp \tab numeric \tab bottom temperature\cr
[,8] \tab depth \tab numeric \tab depth\cr
[,9] \tab lon \tab numeric \tab longitude\cr
[,10] \tab lat \tab numeric \tab latitude\cr
[,11] \tab reg \tab character \tab region name\cr
[,12] \tab datetime \tab c("POSIXct", "POSIXt") \tab date and time\cr
[,13] \tab season \tab character \tab season of year\cr
[,14] \tab year \tab character \tab year\cr
[,15] \tab stratum \tab character \tab stratum, site; based on 1 degree grid\cr
[,16] \tab common \tab character \tab species common name\cr
[,17] \tab species \tab character \tab species name\cr
[,18] \tab nAgg2 \tab integer \tab number of aggregations involved for this species-place-time\cr
[,19] \tab haulid \tab character \tab most specific space-time identifier\cr
[,20] \tab K \tab integer \tab replicate visits to a stratum during the time_lvl period\cr
[,21] \tab Kmax \tab integer \tab maximum number of visits across strata during a given year\cr
}}
\usage{
ebs.a
}
\description{
Aggregated data from Eastern Berring Sea, small subset for testing
}
\seealso{
\code{\link{ebs.agg2}}
}
\keyword{datasets}

