% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ebs.agg2}
\alias{ebs.agg2}
\title{ebs.agg2}
\format{A dim = 254654 x 21 data.table data.frame:
\tabular{rlll}{
[,1] \tab aggStrat \tab character \tab insert_description_here\cr
[,2] \tab time_lvl \tab character \tab insert_description_here\cr
[,3] \tab spp \tab character \tab insert_description_here\cr
[,4] \tab wtcpue \tab numeric \tab insert_description_here\cr
[,5] \tab abund \tab integer \tab insert_description_here\cr
[,6] \tab stemp \tab numeric \tab insert_description_here\cr
[,7] \tab btemp \tab numeric \tab insert_description_here\cr
[,8] \tab depth \tab numeric \tab insert_description_here\cr
[,9] \tab lon \tab numeric \tab insert_description_here\cr
[,10] \tab lat \tab numeric \tab insert_description_here\cr
[,11] \tab reg \tab character \tab insert_description_here\cr
[,12] \tab datetime \tab c("POSIXct", "POSIXt") \tab insert_description_here\cr
[,13] \tab season \tab character \tab insert_description_here\cr
[,14] \tab year \tab character \tab insert_description_here\cr
[,15] \tab stratum \tab character \tab insert_description_here\cr
[,16] \tab common \tab character \tab insert_description_here\cr
[,17] \tab species \tab character \tab insert_description_here\cr
[,18] \tab nAgg2 \tab integer \tab insert_description_here\cr
[,19] \tab haulid \tab character \tab insert_description_here\cr
[,20] \tab K \tab integer \tab insert_description_here\cr
[,21] \tab Kmax \tab integer \tab insert_description_here\cr
}}
\usage{
ebs.agg2
}
\description{
#' Aggregated data from Eastern Berring Sea
}
\seealso{
\code{\link{ebs.a}}
}
\keyword{datasets}

