% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_distance.R
\name{event_distance}
\alias{event_distance}
\title{Event Distance}
\usage{
event_distance(x, positions = seq_along(x), event_value = 1,
  keep_sign = FALSE)
}
\arguments{
\item{x}{vector of values indicating events (and arbitrary values indicating non-events)}

\item{positions}{indicating locations separating distance between values in \code{x}; see example use in \code{\link{event_stretches}} and its use of \code{year}; default assumes event spacing}

\item{event_value}{value in \code{x} inidicating an event}

\item{keep_sign}{Logical; should distances retain their sign? See Note}
}
\description{
Distance from each event to nearest non-event
}
\details{
Finds whether each value in a streak (run) is closer to the beginning or the end of the streak, and then finds the distance to whichever is nearest.
}
\note{
When \code{keep_sign} is TRUE, the sign of values equidistant from 2 non-events may be misleading: even though the event is not closer to either, the sign will be negative.
}

