% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_stretches.R
\name{event_stretches}
\alias{event_stretches}
\title{Find and Classify Unbroken Stretches of an Event}
\usage{
event_stretches(X)
}
\arguments{
\item{X}{a data.table with columns for \code{present}, \code{year}, \code{now_ext}, \code{col}, and \code{ext_dist_sign} (see Details)}
}
\value{
Returns a named list of length 2: \code{stretch_id} and \code{hybrid_part}. 

The values in \code{stretch_id} indicate the following:  
\tabular{ll}{
	-1 \tab the stretch is an extinction-only stretch, which can only happen for a stretch at the beginning of a vector b/c the first element in \code{present} is 1, so there is no initial 0-to-1 transition \cr
	-2 \tab colonization-only stretch; only occurs at end of vector where last value in \code{present} is 1 (no extinction) \cr
	1,2,... \tab a hybrid stretch, occurs for a stretch in \code{present} that is begun and terminated by a 0; this is the most generic case; the integer value serves to distinguish among multiple hybrid stretches \cr
}

When the stretch is classified as a "hybrid", some of the 1's will be closer to the initial 0 than to the terminating 0. Those closer to the inital 0 are considered "post-colonization", whereas those closer to the final 0 are considered "pre-extinction". Proximiting is defined using \code{year}. Ties are ~arbitrarily attributed to post-coloniztion.

The values in \code{hybrid_part} indicate the following:
\tabular{ll}{
	1 \tab post-colonization \cr
	2 \tab pre-extinction \cr
}
}
\description{
Like, run length encoding (see \code{\link{rle}}), this function finds stretches of a vector that contain consecutive repeats of a particular value. This function then procedes by classifying each event in a stretch as being closer to the beginning or end of the stretch.
}
\details{
The contents of the columns in \code{X} are as follows:
\tabular{ll}{
	\code{present} \tab 1 for event, 0 for non-event. "Streaks" refer to consecutive 1's \cr
	\code{year} \tab calendar year, or any time index indicating distance (space, time, etc) between events. \cr
	\code{now_ext} \tab "now extinct"; integer vector of 1 or 0. 1 indicates that \code{present} contains the first 0 following a streak of 1's \cr
	\code{col} \tab "colonization"; integer vector of 1 or 0. 1 indicates that \code{present} contains the first 1 at the beginning of a streak of 1's \cr
	\code{ext_dist_sign} \tab "signed distance to extinction"; output from \code{\link{event_distance}}; magntiude is distance in \code{year}s to non-event, with negative numbers indicating that closest non-event is in past (post-colonization) \cr
}
}

