% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_sample.R
\name{format_range}
\alias{format_range}
\title{Format data to calculate range
Reformat data for use in resampling to estimate range size}
\usage{
format_range(Xdat, sppName)
}
\arguments{
\item{Xdat}{a data.table that is a subset of \code{trawlDiversity::data_all}; subset to one region-year}

\item{sppName}{the name of a species; corresponds to the "spp" column in \code{trawlDiversity::data_all}}
}
\description{
Format data to calculate range
Reformat data for use in resampling to estimate range size
}

