% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ab.R
\name{get_ab}
\alias{get_ab}
\title{Get Alpha and Beta}
\usage{
get_ab(iD, o, yr)
}
\arguments{
\item{iD}{inputData element from output of \code{\link{run_msom}}}

\item{o}{out element from output of \code{\link{run_msom}}}

\item{yr}{the year corresponding to input}
}
\value{
A data.table with iterations of all alpha and beta parameters, which are species-specific. Columns for chain, parameter, value (value is a sample of the posterior), par (alpha or beta), ab_ind (which alpha/ beta parameter, which is the first dimension in parameter), spp_id (species index, second dimension in parameter), spp (species name), and year.
}
\description{
Get alpha and beta parameters given output of an MSOM model run in Stan or JAGS.
}

