% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colonizers.R
\name{get_colonizers}
\alias{get_colonizers}
\title{Get Colonizers}
\usage{
get_colonizers(d, smry = TRUE)
}
\arguments{
\item{d}{An input data.table with columns spp, year (may have others, but they aren't used)}

\item{smry}{Logical, when TRUE (default), the results are summarized in terms of number of species in each category; if FALSE, species-specific categorizations are return}
}
\value{
A data.table with a column for year and columns relating to colonization, extinction, and persistence. If \code{smry} is TRUE, remaining columns are \code{n_col} (number of colonizations), \code{n_ext} (number of extinctions), and \code{n_pers} (number of species persisting). If \code{smry} is FALSE, columns are year, spp, and value. If value is 1 that species colonized (is present now, wasn't last year), if -1 it went extinct (not present now, but was last year), and if 0 it persisted (present both last year and this year).
}
\description{
Identify species colonizing, going extinct, and persisting between years
}

