% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_all.R
\name{get_data_all}
\alias{get_data_all}
\title{Get Data from All Regions}
\usage{
get_data_all(region)
}
\arguments{
\item{region}{Character indicating region abbreviation}
}
\value{
a data.table
}
\description{
Get data from all regions, trimmed and ready for MSOM
}

