% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_iters.R
\name{get_iters}
\alias{get_iters}
\title{Get sampled iterations of posterior}
\usage{
get_iters(X, pars, lang = c("JAGS", "Stan"), ignore_brackets = TRUE)
}
\arguments{
\item{X}{an object containing model output, e.g., as returned by \code{stan} or \code{jags}}

\item{pars}{Character vector of parameter names. Note that naming coventions differ slightly for Stan and JAGS}

\item{lang}{character vector indicating the language that generated X (either "JAGS" or "Stan")}

\item{ignore_brackets}{logical indicating whether or not the brackets in parameter names should be ignored when matching \code{pars} to parameter names in model output. Only implemented for \code{lang == "JAGS"} at the moment, and is only used if \code{pars} does not contain brackets itself.}
}
\value{
A data.table with columns for each of \code{pars} and a column for \code{chain}, the latter of which indicates which of the chains the sample is from.
}
\description{
Given model output from either Stan or JAGS, return a data.table of posterior samples for specified parameters
}
\details{
It is important to remember that models in the two languages might have different parameters/ parameter names. For example, if there is only one \code{beta_mu} parameter, it is \code{beta_mu} in JAGS and \code{beta_mu[1]} in Stan. Also, \code{logit_psi} and \code{logit_theta} are tracked in Stan but not JAGS, and \code{Z} parameters are only in JAGS. Both the way the models are written and the languages themselves contribute to these differences.
}

