% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_show_up_spp.R
\name{get_show_up_spp}
\alias{get_show_up_spp}
\title{Get species that show up and remain}
\usage{
get_show_up_spp(X, n_yrs_start = 2, n_yrs_remain = 5,
  n_yrs_miss_after = 1)
}
\arguments{
\item{X}{A data.table containing trawl data}

\item{n_yrs_start}{after this many years, species seen for the first time "show up"; acts as a burn-in}

\item{n_yrs_remain}{how many years must be remaining in the time series after a species "shows up" to evaluate it as "sticking around"}

\item{n_yrs_miss_after}{how many years the species can be missing during remaining (post show-up) years and still be considered to "stick around"}
}
\value{
A data.table containing the region, year of appearance, species, stats for how long and how consistently that species showed up.
}
\description{
A crude method for determining which species show up in the data set due to enhanced detection
}
\details{
Function looks for species that appear after the second year, but before the final 5 years, and that in the remaining years (of which there must be at least 5) the species is observed in at least every year but 1. So a species that shows up after many years of sampling, and then is seen every year thereafter, will be returned by this function. Those species might be showing up and sticking around b/c of changing detection by the survey (I'd expect species to come on the scene more slowly, such that after it first shows up, there should still be some years it isn't observed)
}

