% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_rich_sample.R
\name{local_rich_sample}
\alias{local_rich_sample}
\title{Resample Observed Local Richness}
\usage{
local_rich_sample(X)
}
\arguments{
\item{X}{a data.table that is a subset of \code{\link{trawlDiversity::data_all}}. Subset to a region-year.}
}
\value{
returns a numeric value of length 1, the average local richness
}
\description{
The number of tows affects richness, this function calculates site-level richness after rarefying tow intensity
}
\details{
Value returned is the rarefied (to 1 tow) cross-site average richness. Each site in the region is expected, on average, to have this many species observed in it if 1 tow is made per site.
}
\seealso{
\code{\link{range_sample}}
}

