% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{localAC}
\alias{localAC}
\title{Results for local autocorrelation}
\format{list data.table}
\usage{
localAC
}
\description{
Includes distance, neighborhood, and local autocorrelation (and associated pvalues, z scores, lon lat) for each site
}
\seealso{
\code{\link{spatial_ac}}
}
\keyword{datasets}

