% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mapDat.R
\name{make_mapDat}
\alias{make_mapDat}
\title{Make data for maps related to diversity analysis}
\usage{
make_mapDat(p)
}
\arguments{
\item{p}{The \code{p} data object from the process_msom_script.R file.}
}
\value{
a data.table
}
\description{
For each region, get the data appropriate for maps of richness and environmental metrics.
}
\details{
The data.table contains summary statistics to collapse time (long-term mean, e.g.), and retains stratum-specific information.
}

