% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_owin.R
\name{make_owin}
\alias{make_owin}
\title{Make owin from Region Outline}
\usage{
make_owin(X, outlines)
}
\arguments{
\item{X}{a data.table of mapDat, e.g.}

\item{outlines}{region outlines, e.g., the outlines object}
}
\description{
Make an owin object for spatstat package
}
\examples{
trawlDiversity::make_owin(mapDat, outlines)
}

