% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mapDat}
\alias{mapDat}
\title{Data for Maps}
\format{data.table}
\usage{
mapDat
}
\description{
Spatial information related to colonization, extinction, richness, depth, and bottom temperature; does not contain any explicit temporal information
}
\details{
Meant for making maps. Other processed data sets emphasize the temporal dimension. This data set collapses time, and emphasizes 2D spatial information.

\tabular{lll}{
\code{stratum} \tab character \tab sampling stratum, redefined via \code{\link{trim_msom}} in \code{\link{get_data_all}} (and therefore in \code{\link{data_all}}) \cr
\code{lon} \tab numeric \tab longitude in degrees East (such that negative values are in the Western hemisphere) \cr
\code{lat} \tab numeric \tab latitude \cr
\code{depth} \tab numeric \tab depth in meters; in each year hauls in the stratum are averaged, then this value is the cross-year average of those annual depths \cr
\code{n_spp_col_weighted, n_spp_ext_weighted} \tab numeric \tab total colonizers per stratum colonized (# species colonizing this stratum / # of strata colonized by each species) in this stratum, averaged across years; similar for ext= extinction \cr
\code{reg} \tab character \tab region \cr
\code{avgRich, sdRich} \tab numeric \tab cross-year average (standard deviation) richness in this stratum, based on naive richness (not MSOM estimate) \cr
\code{avgBtemp, sdBtemp} \tab numeric \tab cross-year average (standard deviation) of bottom temperature in this stratum \cr
}
}
\seealso{
\code{\link{spp_master}}, \code{\link{comm_master}}
}
\keyword{datasets}

