% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mapOwin}
\alias{mapOwin}
\title{Region outlines in owin format}
\format{owin}
\usage{
mapOwin
}
\description{
Transformed \code{\link{outlines}} for use with the \code{spatstat package}
}
\seealso{
\code{\link{make_owin}}
}
\keyword{datasets}

