% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msom.R
\name{msom}
\alias{msom}
\title{Run MSOM}
\usage{
msom(X, method)
}
\arguments{
\item{X}{Properly formatted MSOM data}

\item{method}{Character specifying MSOM model to run}
}
\description{
Run a multispecies occupancy model on properly formatted trawl data. Can access different MSOM structures through this function.
}
\details{
In development, coming soon.
}

