% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msom_params.R
\name{msom_params}
\alias{msom_params}
\title{Print Names of MSOM Parameters}
\usage{
msom_params(language, model_type, nT, Jmax, nU, nV, nS, noIndex = FALSE)
}
\arguments{
\item{language}{The language of the MSOM model (JAGS or Stan)}

\item{model_type}{The type of model (Dynamic or Static)}

\item{nT}{the number of years}

\item{Jmax}{the total number of unique sites ever visited}

\item{nU}{the number of presence covariates}

\item{nV}{the number of detection covariates}

\item{nS}{the number of species (including 0's)}

\item{noIndex}{logical, default FALSE; if TRUE, strip indices from parameter names and return unique values (this can return a much smaller object, and greatly reduce computation time)}
}
\value{
A named list with the following elements:
\tabular{ll}{
n \tab a named vector with integer elements (see below)\cr
params \tab the names of all potential parameters\cr
params_main \tab the names of all main-effect parameters\cr
params_random \tab the names of all random-effect parameters\cr
params_latent \tab the names of all latent paramters\cr
}

\code{n} contains the following elements
\tabular{ll}{
nP \tab total parameters\cr
n_main \tab total main-effect parameters, including hyperparameters\cr
n_random \tab number of random-effect parameters\cr
n_latent \tab number of random-effect parameters\cr
}
}
\description{
Given a specified language and model type (as in \code{\link{run_msom}}), and the dimensions of the data (e.g., \code{\link{msomData}}), list the potential parameter names as well as the total number of parameters
}
\details{
Currently does not include reporting covariates specified as random variables, nor do these extra parameters factor into the total parameter count. Z is only available in JAGS, as is w. JAGS is parameterized with tau's, Stan with sd's. Only dynamic models include parameters for persistence (phi) and colonization (gamma). Even though they aren't strictly parameters, all models and languages can report parameters for probability of presence (Psi) and probability of detection (Theta). Most hyperparameters are available reporting; however, parameters used strictly to generate non-centered distrubtions (used in Stan only) are not available. In addition to those already mentioned, parameters available include the following (and the pertinent variations on these parameters): alpha, beta, Omega.

Parameter names are semi-redundant: e.g., the 'w' parameter in a JAGS model is species-specific. In the output list of all possible parameters, 'w' will be listed, as will 'w[1]', 'w[2]', ..., 'w[nS]'. The interpretation here is that 'w' implies all of its indices, but that both languages accept both specifications, so both versions are included. Note that this introduces a small asymmetry between the lengths reported in \code{n} and the lengths in the other elements of the output list (see Examples).
}
\examples{
lang <- "JAGS"
mt <- "Static"
out <- msom_params(lang, mt, nT=3, Jmax=4, nU=3, nV=1, nS=5)

# In reference to the asymetry mentioned in Details
out[["n"]][2:4] == sapply(out[3:5], length) # symmetrical
out[["n"]][1] < length(out[["params"]]) # asymmetric
}

