% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{outlines}
\alias{outlines}
\title{Regional Polygon Outlines}
\format{data.table}
\usage{
outlines
}
\description{
Like a convex hull, but a bit more form-fitting. They should all be counterclockwise (to work with \code{spatstat::ppp}), and they are not quite closed. These were manually created.
}
\details{
Meant for making maps. Other processed data sets emphasize the temporal dimension. This data set collapses time, and emphasizes 2D spatial information.

\tabular{lll}{
\code{reg} \tab character \tab region \cr
\code{lonP} \tab numeric \tab longitude in degrees East (such that negative values are in the Western hemisphere) \cr
\code{latP} \tab numeric \tab latitude \cr
}
}
\keyword{datasets}

