% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ab.R
\name{plot_ab}
\alias{plot_ab}
\title{Plot Alpha and Beta}
\usage{
plot_ab(X, t_spp, plt_img = TRUE, plt_pts = TRUE, ...)
}
\arguments{
\item{X}{a data.table with columns for year and value}

\item{t_spp}{character indicating a species name}

\item{plt_img}{logical indicating whether or not to plot a picture of the species in the background}

\item{plt_pts}{logical indicating whether or not to plot the posterior points, or just the mean of the posterior}

\item{...}{arguments passed to both \code{sppImg} and \code{plot}}
}
\value{
the output of \code{sppImg}, or if \code{plt_img} is \code{FALSE}, returns \code{NULL}
}
\description{
Plots a species image and the parameter in a provided data.table
}
\details{
Used in process_msom_figures.R. X is assumed to have been subset to the specific species and the desired annual posterior for value, such as when value is a species-specific paraemter for alpha or beta.

The function tries to adjust the opacity of the points according to the size of the panel, with the intention of not completely obscuring the background image and so that opacity can be used as an indicator of posterior density. However, if the panel is too small, the transparency may be truncated to 0, so the posterior dots are just invisible.
}

