% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_btemp_map.R
\name{plot_btemp_map}
\alias{plot_btemp_map}
\title{Plot Map of Bottom Temperature}
\usage{
plot_btemp_map(prn, Figures, FUN = "dev.new", ...)
}
\arguments{
\item{prn}{the p object (processed msom; output from \code{process_msomStatic})}

\item{Figures}{option list to which the figure and its information should be added}
}
\value{
Returns the Figure object
}
\description{
Give the processed MSOM, plot a map of bottom temperature in each year for the region
}
\details{
The function uses \code{unpack_p} to get much of the information it needs. The \code{Figures} object is a list whose first level is intended to be the region. The second level is specific to each figure. The third level has 3 elements: 'figure', 'name', and 'dim'. The 'figure' element is the result of using \code{\link{recordPlot}} on what is plotted. The name of the figure is, e.g., what the saved figure would be called. The dim is the width and height (in that order) in inches.
}
\seealso{
There are several functions that are run through the process_msom_figures script. Richness and temperature plots are \code{\link{plot_btemp_map}}, \code{\link{plot_rich_bt_scatter}}, and \code{\link{plot_rich_bt_ts}}. Figures for colonization, extinction, and the species and places associated with those processes are \code{\link{plot_ce_wrap}}, \code{\link{plot_col_vs_unobsSpp}}, \code{\link{plot_colExt_perStrat}}, and \code{\link{plot_rank_temp}}. Figures for diagnostics are \code{\link{plot_traceplot}} and \code{\link{plot_post_corr}}.
}

