% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce.R
\name{plot_ce}
\alias{plot_ce}
\title{Plot a Species that Colonizes and/or goes Extinct}
\usage{
plot_ce(t_sco, pad_top_mar = 2, plt_pts = FALSE, ...)
}
\arguments{
\item{t_sco}{a character indicating a species name}

\item{pad_top_mar}{numeric indicating the amount to increase the size of the top margin for the first of the 3 panels}

\item{plt_pts}{logical indicating whether or not to plot points for first figure (see \code{plot_ab})}

\item{...}{arguments passed to \code{plot_ab}, \code{plot}, and \code{plot_space}}
}
\value{
nothing
}
\description{
Plots 3 panels for a given species: its image and time series of detectability, the location of colonizations, and time series of biomass
}
\details{
Used in process_msom_figures.R
}

