% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce_setup.R
\name{plot_ce_setup}
\alias{plot_ce_setup}
\title{Setup Colonization & Extinction Plot}
\usage{
plot_ce_setup(spp2use, width.max = 12, height.max = 7,
  max_spp_columns = 15, nPlots = 3)
}
\arguments{
\item{spp2use}{character vector of species names}

\item{width.max}{the maximum width of the figure in inches}

\item{height.max}{the maximum height of the figure in inches}

\item{max_spp_columns}{the maximum number of columns to use when arranging panels}

\item{nPlots}{the number of plots per species; determined by \code{\link{plot_ce}}}
}
\value{
Returns a numeric vector of length 2 indicating the width and height of the output figure, in inches
}
\description{
Creates layout and dimensions for c/e plot
}
\details{
Shouldn't need to use this function directly. See \code{\link{plot_ce_wrap}}
}
\seealso{
There are several functions that are run through the process_msom_figures script. Richness and temperature plots are \code{\link{plot_btemp_map}}, \code{\link{plot_rich_bt_scatter}}, and \code{\link{plot_rich_bt_ts}}. Figures for colonization, extinction, and the species and places associated with those processes are \code{\link{plot_ce_wrap}}, \code{\link{plot_col_vs_unobs}}, \code{\link{plot_colExt_perStrat}}, and \code{\link{plot_rank_temp}}. Figures for diagnostics are \code{\link{plot_traceplot}} and \code{\link{plot_post_corr}}.
}

