% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colExt_perStrat.R
\name{plot_colExt_perStrat}
\alias{plot_colExt_perStrat}
\title{Number of Colonizations and Extinctions in Each Stratum}
\usage{
plot_colExt_perStrat(prn, Figures, FUN = "dev.new", ...)
}
\arguments{
\item{prn}{the p object (processed msom; output from \code{process_msomStatic})}

\item{Figures}{option list to which the figure and its information should be added}

\item{FUN}{graphical device function}

\item{...}{arguments passed to plot_device}
}
\value{
Returns the Figure object
}
\description{
Plots the number of colonizations, extinctions, and their difference as the time series total in each stratum
}
\details{
Each of the 3 types of plots has two panels: The first panel is a map of the strata indicating the the values seen in the data. Each stratum is a point. The second panel is a thin plat spline of the first panel (see \code{fields::Tps}), which acts to smooth the values in the convex hull of the strata.

The units of the map are a bit complicated. They generally indicate the mean number of colonizations (or extinctions, or C-E) per year for each stratum. However, the number of 'colonizations' for a stratum in a given year is the number of species that colonized the region at that stratum, divided by the total number of strata that those same species colonized that year. Thus, species that colonize many strata do not contribute much to the colonization score of any single stratum, and the sum of each stratum's colonization score for a year is the number of species that colonized the whole region. It is this colonization score that is averaged over years (which can be different from the sum across years b/c some strata aren't sampled every year).

The function uses \code{unpack_p} to get much of the information it needs. The \code{Figures} object is a list whose first level is intended to be the region. The second level is specific to each figure. The third level has 3 elements: 'figure', 'name', and 'dim'. The 'figure' element is the result of using \code{\link{recordPlot}} on what is plotted. The name of the figure is, e.g., what the saved figure would be called. The dim is the width and height (in that order) in inches.
}
\seealso{
There are several functions that are run through the process_msom_figures script. Richness and temperature plots are \code{\link{plot_btemp_map}}, \code{\link{plot_rich_bt_scatter}}, and \code{\link{plot_rich_bt_ts}}. Figures for colonization, extinction, and the species and places associated with those processes are \code{\link{plot_ce_wrap}}, \code{\link{plot_col_vs_unobsSpp}}, \code{\link{plot_colExt_perStrat}}, and \code{\link{plot_rank_temp}}. Figures for diagnostics are \code{\link{plot_traceplot}} and \code{\link{plot_post_corr}}.
}

