% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_device.R
\name{plot_device}
\alias{plot_device}
\title{Plot Device}
\usage{
plot_device(x, FUN, path = "trawlDiversity/pkgBuild/figures/pmfPieces", ...)
}
\arguments{
\item{x}{a Figures list}

\item{FUN}{a character indicating a graphical device function}

\item{path}{folder where the figure will be saved, if FUN is not \code{dev.new}}

\item{...}{other arguments passed to \code{FUN}}
}
\description{
Prepare the graphical device for use with processed msom figures
}
\details{
Used in the functions called in the plotting functions used in process_msom_figures.R
}

