% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_space.R
\name{plot_space}
\alias{plot_space}
\title{Plot Spatial Data}
\usage{
plot_space(x, y, z, scatter = FALSE, ...)
}
\arguments{
\item{x}{x values (independent)}

\item{y}{y values (independent)}

\item{z}{z values (dependent value)}

\item{scatter}{logical, if true, plots a scatter plot of raw values; else, smooths data with Krigging}

\item{...}{arguments passed to plot via colScat}
}
\description{
Plot spatial data consistent of X, Y, and Z components
}

