% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsTau.R
\name{post_trend}
\alias{post_trend}
\title{Trend through Posterior Distributions}
\usage{
post_trend(x, Y, nSamp = 100)
}
\arguments{
\item{x}{time step}

\item{Y}{matrix of posterior samples. Each column is a time step, each row is a sample of the posterior}

\item{nSamp}{number of times each set of posterior samples should be resampled to fit trend. I.e., number of bootstrap iterations.}
}
\value{
A named numeric vector of length 3, containing Kendall's Tau, the associated Z statistic, and the p-value
}
\description{
Fit a trend to a time series of posterior estimates. Ends up being similar to bootstrapping.
}
\details{
Uses Kendall's Tau and accounts for serial correlation in the time series (up to ARMA(2,2)). Currently assumes that there are no ties in the resampled Y. If there are, the p-value may be incorrect.
}
\examples{
n <- 20
x <- 1:n
slope <- 0.5
n_post <- 50
ar_coeff <- 0.85
y <- c(slope*x+arima.sim(model=list(ar=c(ar_coeff)), n=n))
y_post <- sapply(y, function(x)rnorm(n_post, mean=x, sd=1))
post_trend(x, y_post, 10)

}
\seealso{
\code{\link{tsTau}}
}

