% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_msomStatic.R
\name{process_msomStatic}
\alias{process_msomStatic}
\title{Process Output from msomStatic (annual)}
\usage{
process_msomStatic(rm_out, reg)
}
\arguments{
\item{rm_out}{A list with length equal to number of years in a region, with each element containing output from run_msom}

\item{reg}{Region name (character)}
}
\description{
Processes a list structured as run_msom output (level 2) within a year (level 1) to be used to summarize diversity. Note that this function then only processes 1 region at a time, so it expects a list whose length is equal to the number of years for a region. Currently, intended to work with Stan model.
}
\details{
Right now only intended for use with specific structuring of the output, so that it matches the output expected from running each year separately using the Stan version of the msomStatic model.
}

