% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_obsRich.R
\name{process_obsRich}
\alias{process_obsRich}
\title{Process Observed Data for Richness}
\usage{
process_obsRich(X, msom_yrs)
}
\arguments{
\item{X}{a data.table containing data for a region. Should \code{\link{data_all}} (1 year of which used as input to \code{\link{run_msom}})}

\item{msom_yrs}{Optional vector of integers indicating calendar years analyzed by the msom, and that will be used to subset \code{X}}
}
\description{
Used observations, as opposed to model output, to summarize elements related to regional richness
}
\details{
The analysis performed by this function was previously contained in process_msomStatic.
}

