% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mk_cov_rv.R
\name{prop_sd_pow}
\alias{prop_sd_pow}
\title{Propagate standard deviation from raising to power}
\usage{
prop_sd_pow(A, sigma_A, b = 2, a = 1)
}
\arguments{
\item{A}{The random variables being operated on}

\item{sigma_A}{Standard deviations of A}

\item{b}{The exponent}

\item{a}{the multiplication coefficient}
}
\description{
Calculates the standard deviation of a random variable created by multiplying another RV by a constant and raising it to a power
}
\details{
Of the form f = a*A^b. Sigma must be provided. Elements of A are treated as potentially separate RV's.
}

