% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_metrics.R
\name{response_metrics}
\alias{get_response}
\alias{psi.max}
\alias{psi.opt}
\alias{psi.tol}
\alias{response_metrics}
\title{Response Metrics for Quadratic Model}
\usage{
get_response(alphas, X, n_samp = NULL, n_grid = NULL)

psi.opt(b1, b2)

psi.tol(b2)

psi.max(b0, b1, b2)
}
\arguments{
\item{alphas}{a data.table with alpha values; rows are posterior iterations, a1,a2,a3,a4,a5 are columns (see details)}

\item{X}{a data.table with columns for \code{bt} (bottom temperature) and \code{depth}; bt associated with a2 and a3, depth with a4 and a5}

\item{n_samp}{integer indicating (random) subsample of posterior to use; if NULL (default), no subsampling is performed}

\item{n_grid}{number of depth and temperature values over which to estimate response; if NULL (default), response is calculated only for observed combinations in \code{X}}

\item{b1}{a linear coefficient value (like a2 or a4 in \code{alphas})}

\item{b2}{a quadratic coefficient value (like a3 or a5 in \code{alphas})}

\item{b0}{an intercept parameter value (like a1 in \code{alphas})}
}
\description{
For quadratic models predicting responses across an environmental gradients, extract the response and metrics (optimum value, e.g.)

\code{get_response}: Get model predicted response across a range of depth and temperature values

\code{psi.opt}: Optimum environmental value

\code{psi.tol}: Tolerance across environmental values

\code{psi.max}: Response achieved at environmental optimum
}
\details{
The columns in \code{alphas} are parameters estimated from the MSOM model. The MSOM model has a form of \code{y=a1*1 + a2*bt+a3*bt^2 + a4*depth+a5*depth^2}. Thus, the columns in alpha are interpreted as follows:
\tabular{ll}{
\code{a1} \tab an intercept term, corresponds to \code{b0} \cr
\code{a2} \tab a linear coefficient associated with bottom temeprature, corresponds to \code{b1} \cr
\code{a3} \tab a quadratic coefficient associated with bottom temperature, corresponds to \code{b2} \cr
\code{a4} \tab a linear coefficient associated with depth, corresponds to \code{b1} \cr
\code{a5} \tab a quadratic coefficient assoicated with depth, corresponds to \code{b2} \cr
}
where \code{b0}, \code{b1}, \code{b2} are arguments to the response metric functions (\code{psi.opt}, \code{psi.tol}, and \code{psi.max}). Note, then, that the response metric functions do not consider models with 2 predictors, but only 1, hence the different notation.
}

