% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_msom.R
\name{run_msom}
\alias{run_msom}
\title{Run an MSOM Model}
\usage{
run_msom(reg = c("ai", "ebs", "gmex", "goa", "neus", "newf", "ngulf", "sa",
  "sgulf", "shelf", "wcann", "wctri"), params_out = c("params", "params_main",
  "params_random", "params_latent", "custom"), custom_params = NULL,
  model_type = c("Dynamic", "Static"), chains = 4,
  cores = parallel::detectCores()/2, iter, language = c("JAGS", "Stan"),
  test = FALSE, test_sub = list(stratum = 4, year = 3, spp = 10),
  seed = 1337)
}
\arguments{
\item{reg}{character, region name}

\item{params_out}{character vector specifying category of parameters to report in output; if custom, points to \code{custom_params} as well}

\item{custom_params}{possible characters specifying custom parameters, potentially indexed; see Details}

\item{model_type}{type of model. Currently both are multiyear. 'Dynamic' includes terms for persistence and colonization (species-specific), whereas 'Static' simply 'stacks' the years.}

\item{chains}{integer number of chains, default is 4}

\item{cores}{integer number of cores for parallel processes; default is half of avaialble cores}

\item{iter}{number of iterations; the default depends on \code{language} and on /code{test}; testing in Stan is 50, non-testing is 200. Testing in JAGS is 500, non-testing is 5000.}

\item{language}{character indicating the language to be used --- JAGS or Stan}

\item{test}{Logical, whether to do this run as a 'test' run. Default is FALSE. If TRUE, fewer iterations are run (unless overridden by non-default), and the data set is subsetted}

\item{test_sub}{a named list with elements 'stratum', 'year', and 'spp'. Each element should be an integer indicating the number of levels to select for each of those dimensions. Used for subsetting when \code{test} is TRUE.}

\item{seed}{integer random number seed}
}
\description{
Run a multispecies occupancy model on trawl data. The model is multi-year, and there are two basic versions; first one is a 'static' model for which no temporal process is explicitly modelled (the years are 'stacked'). The second is a 'dynamic' model that includes persistence and colonization parameters as processes that facilitate the state transition between years. Each of these models can be run in either JAGS or Stan. Both models accept covariates for the detection and presence processes, and both sets of covariates can be modelled as random variables, constants (varying, but known precisely), or a mixture of the two.
}
\details{
Both \code{params_out} and \code{custom_params} must find matches in the output of \code{\link{msom_params}}. For all parameters, use 'params'; main-effect parameters specified via 'params_main'; random-effect parameters via 'params_random'. Latent stochastic nodes/ parameters via 'params_latent'. Additional flexibility offered by specifying 'custom', which will add manually specified parameters from \code{custom_params}.
}

