% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sac.R
\name{run_sac}
\alias{run_sac}
\title{Run Species Accumulation Curve Estimate of Richness}
\usage{
run_sac(regs)
}
\arguments{
\item{regs}{Character vector of region names}
}
\description{
Calculate richness using a species accumulation curve
}

