% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqAgg.R
\name{seqAgg}
\alias{seqAgg}
\title{Sequential Aggregation}
\usage{
seqAgg(X, val, FUN, by, ...)
}
\arguments{
\item{X}{a data.table}

\item{val}{character vector indicating columns in X that should be sequuentially aggregated}

\item{FUN}{a list of functions}

\item{by}{a list of character vectors; should be of same length as FUN}

\item{...}{arguments to be passed to each element of \code{FUN}}
}
\description{
Sequentially aggregate columns in a data.table, varying the function and 'by' values
}
\details{
The lengths of \code{FUN} and \code{by} should be of the same length, but the length of \code{val} is independent.

Useful when a summary statistic needs to be computed sequentially (possibly taking the mean over some dimensions before computing the standard deviation, e.g.)
}
\examples{
neus_dat <- data_all[reg=="neus"&year==2010]
a1 <- c("K","stratum")
a2 <- c("stratum")
a3 <- c("")
a_list <- list(a1, a2, a3)
f_list <- list(mean, mean, sd)

neus_dat[,sd(btemp,na.rm=TRUE)] # naive
seqAgg(neus_dat, val="btemp", FUN=f_list, by=a_list, na.rm=TRUE) # sequential
seqAgg(neus_dat, val=c("btemp","depth"), FUN=f_list, by=a_list, na.rm=TRUE) # 2 columns

}

