% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{show_up_spp}
\alias{show_up_spp}
\title{Species that have shown up late, and stayed around}
\format{data.table}
\usage{
show_up_spp
}
\description{
Indicates species that are suspected of showing up in data set due to enhanced identification, not because of actual colonization of the region
}
\keyword{datasets}

