% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spp_master}
\alias{spp_master}
\title{Master Species-Level Data}
\format{data.table}
\usage{
spp_master
}
\description{
Master data set of species time series related to colonization, extinction, richness, MSOM results, and environmental values.
}
\details{
\tabular{lll}{
\code{reg} \tab character \tab region code \cr
\code{spp} \tab character \tab species scientific name \cr
\code{year} \tab integer \tab sampling year \cr
\code{detect_mu} \tab numeric \tab mean of posterior of species-specific detection parameter for the current year-model \cr
\code{detect_sd} \tab numeric \tab standard deviation of posterior of species-specific detection parameter \cr
\code{col} \tab integer \tab 0 or 1, 1 indicates that the species colonized the region this year (was absent in previou sampling year) \cr
\code{ext} \tab integer \tab 0 or 1, 1 indicates that the species will be absent in following sampling year \cr
\code{present} \tab integer \tab 0 or 1, 1 indicates that the species is present in the region in current year \cr
\code{now_ext} \tab integer \tab 0 or 1, 1 indicates that the species is absent this year, but was present in previous sampling year \cr
\code{cumm_yrs_pres} \tab integer \tab number of sampling years that the species has been seen up to (and including) this year \cr
\code{consec_yrs_pres} \tab integer \tab number of consecutive sampling years that the species has been observed \cr
\code{yrs_since_1st} \tab integer \tab number of sampling events (including current) since species was first observed \cr
\code{ext_dist, ext_dist_sign, ext_dist_samp} \tab integer (possibly numeric) \tab see \code{\link{event_distance}}; \code{_samp} is sampling years elapsed, others are actual years elapsed \cr
\code{propStrata} \tab numeric \tab proportion of strata that the species was observed to occupy in current sampling year \cr
\code{bt_ann} \tab numeric \tab cross-stratum average of bottom temperature in current sampling year; not a species-specific value (is same for all species in that region-year) \cr
\code{has_stretches} \tab logical \tab indicates whether or not this spp-year-reg is part of a stretch of presences that is flanked (on at least 1 side) by an absence \cr
\code{stretch_id, hybrid_part} \tab integer \tab see \code{\link{event_stretches}} \cr
\code{stretch_type} \tab character \tab "post_col" for post-colonization stretch, "pre_ext" for pre-extinction stretch \cr
\code{event_year} \tab integer \tab the year of the closest absence, all years in the same stretch have the same \code{event_year} \cr
\code{stretch_length} \tab integer \tab number of sampling years included in the current stretch \cr
\code{bt_opt, bt_tol, prob_max} \tab numeric \tab optimum bottom temperature and bottom temperature tolerance; presence probability at optimum (see \code{\link{response_metrics}}) \cr
\code{depth_base_opt} \tab numeric \tab median depth of strata that are in the top 95\% of strata ranked by estimated probabilities of occurrence for the species; needed so that \code{prob_max} can be calculated, as depth needs to be part of the intercept \cr
\code{btemp_ODS} \tab numeric \tab ODS is optimal depth strata; average bottom temperature in the strata used in \code{depth_base_opt} \cr
\code{nODS} \tab integer \tab number of ODS \cr
\code{bt_opt_avg, bt_tol_avg, detect_mu_avg} \tab numeric \tab cross-year average of this species' (in this region) \code{bt_opt, bt_tol, detect_mu} \cr
\code{ce_categ} \tab character \tab colonization/extinction category; neither means present every year, colonizer means colonized only, leaver means went extinct only, both means has both types of events \cr
}
}
\seealso{
\code{\link{comm_master}}, \code{\link{mapDat}}
}
\keyword{datasets}

