% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streak_length.R
\name{streak_length}
\alias{streak_length}
\title{Streak Length}
\usage{
streak_length(x, streak_val = 1, fill_val = 0)
}
\arguments{
\item{x}{vector of numbers}

\item{streak_val}{value in \code{x} considered as part of streak}

\item{fill_val}{value in output to indicate indices that are part of non-streak}
}
\value{
Returns a vector the same length as \code{x} that contains streak lengths and 0's
}
\description{
Finds streaks in a vector containing consecutively-repeated values, and in each streak indicates the number of values into the streak so far
}
\details{
Functions like \code{\link{rle}}, but within each run (streak) indicates the current position of each value where the first value in the streak is 1 and the last value in the streak is the length of the streak
}
\examples{
streak_length(c(1,1,1,0,0,1,0,1,0,0,1,1))

}

