% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSlope.R
\name{timeSlope}
\alias{timeSlope}
\title{Time Slope}
\usage{
timeSlope(x, y, ...)
}
\arguments{
\item{x}{temporal predictor variable}

\item{y}{response variable}

\item{...}{others arguments; not used}
}
\description{
Calculate the temporal trendline (quickly)
}
\details{
Requires at least 3 non-na values. 

For benchmarking, see: https://github.com/rBatt/trawl/commit/a04d8e166d7093538c698368d48542a0e60bd16c
}

