% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-trawlDiversity.R
\docType{package}
\name{trawlDiversity}
\alias{trawlDiversity}
\alias{trawlDiversity-package}
\title{trawlDiversity: A package for calculating species richness from bottom trawl surveys}
\description{
The package is associated with a study aimed at understanding long-term changes in species richness in marine ecosystems around the North American coastline. It employs multispecies occupancy models (MSOM) to estimate richness, as well as a few other methods. Many of the functions are aimed at visualizing data, and conducting supporting statistical analyses to understand the trends reported from MSOM posterior results (MSOM's are Bayesian models).
}

