% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_msom.R
\name{trim_msom}
\alias{trim_msom}
\title{Trim trawl data for msom}
\usage{
trim_msom(reg, gridSize = 1, grid_stratum = TRUE, depthStratum = NULL,
  tolFraction = 1/3, plot = FALSE, cull_show_up = FALSE)
}
\arguments{
\item{reg}{Character name of region or data.table to be passed to \code{\link{trawlTrim}}.}

\item{gridSize}{numeric grid size to be passed to \code{\link{ll2strat}}}

\item{grid_stratum}{logical, default TRUE; whether or not the strata should be defined on a grid, or used original stratum definition}

\item{depthStratum}{The depth range to use to define stratum; the default does not use depth at all. If using, a reasonable value might be 100.}

\item{tolFraction}{The fraction of years that a stratum can not be sampled yet still be included in output. Default is 1/3, indicating that a stratum will be included in output so long as it is sampled for at least 2/3 of years. Actual number of years tolerated is rounded up. This value determines the strat_tol argument in \code{\link{check_strat}}}

\item{plot}{Logical, default FALSE; for \code{check_strat}, should the stratum tolerance be plotted?}

\item{cull_show_up}{Logical, default \code{FALSE}; should spp in \code{\link{show_up_spp}} be removed when trimming?}
}
\value{
A data.table
}
\description{
Performs the \code{trawlTrim} function the clean region data, then trims down to species (not genera) that have been observed at least 10 times. Does aggregating within a haul/ spp using \code{trawlAgg}. Adds an abundance column, which is actually just 0 if wtcpue == 0, and 1 if wtcpue > 0.
}

