% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_p.R
\name{unpack_p}
\alias{unpack_p}
\title{Unpack P}
\usage{
unpack_p(p_rn)
}
\arguments{
\item{p_rn}{The p object subset to a region}
}
\description{
Unpacks the object containing the processed MSOM results
}
\details{
Make sure to only supply p after it's been subset (e.g., \code{p[[1]]}). The structure of p is such that the first level is the regions, the second is the years, then the third contains the varius elements associated with the processed output of a single model run.
}

