% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale Parameters}
\usage{
unscale(a1, a2, a3, a4, a5, mu1, mu2, sigma1, sigma2)
}
\arguments{
\item{a1}{intercept term}

\item{a2}{coefficient for first predictor}

\item{a3}{coefficient for first predictor squared}

\item{a4}{cofficient for second predictor}

\item{a5}{coefficient for second predictor squared}

\item{mu1}{mean used to scale first predictor}

\item{mu2}{mean used to scale second predictor}

\item{sigma1}{standard deviation used to scale first predictor}

\item{sigma2}{standard deviation used to scale second predictor}
}
\value{
Either a vector (if a1 is of length 1) or a matrix (if a1 is longer than 1) of unscaled parameters.
}
\description{
Unscale parameters from a polynomial multiple linear regression
}
\details{
Assume a model of the general form \code{y ~ a1 + a2*x1 + a3*x1^2 + a4*x2 + a5*x2^2}. Scaling is assumed to be \code{scale(x1)}, e.g. All arguments can be vectors, for which each element should refer to a different model/ regression.

Created to unscale parameters used in an MSOM. In this case, each iteration on each chain effectively referes to a different set of parameters. So if the posterior was 100 iterations sampled, the 'a' arguments would be of length 100, and mu and sigma arguments could be length 1 (and recycled), or have the same value repeated 100 times. Alterntively, if the 'a' arguments refer to models for which the predictors were scaled differently, just make sure that the mu and sigma values are repeated to match to the corresponding models. No checking is done, but this should be fairly intuitive: the coefficients should match up with the scaling constants used to scale the predictors used in the model that estimated the coefficients.
}

