% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colonizers.R
\name{get_colonizers}
\alias{get_colonizers}
\title{Get Colonizers}
\usage{
get_colonizers(d)
}
\arguments{
\item{d}{An input data.table with columns spp, year (may have others, but they aren't used)}
}
\value{
A list of length 5:
\tabular{ll}{
\code{col_dt} \tab colonization events for each year-stratum-species. \code{col_logic} is specific to year-stratum-spp, \code{n_spp_col} is specific to year-stratum (number of species colonizing), and \code{n_strat_col} is specific to year-spp (number of strata occupied by that species in that year [if it colonized the region that year]). Like the weighted version, \code{n_spp_col_weighted} indicates the number of species colonizing a stratum in a year, but if a species colonizes 5 strata, it only contributes 0.2 to each, instead of 1. The number of years a stratum was sampled is \code{yrs_sampled}.
\code{col_ext_dt} \tab colonization (1), extinction (-1), or no change for each species-year \cr
\code{n_cep} \tab the number of colonizations, extinctions, or no-change events in the whole region for each year \cr
\code{n_spp_col_weighted} \tab appropriate aggregation of the column by this name in \code{col_dt} \cr
\code{n_spp_col_weighted_tot} \tab like \code{n_spp_col_weighted}, but summed up across years (and divided by \code{yrs_sampled}) \cr
}
}
\description{
Identify species colonizing, going extinct, and persisting between years
}

