% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mk_cov_rv.R
\name{mk_cov_rv}
\alias{mk_cov_rv}
\alias{mk_cov_rv_pow}
\title{Make Covariate (and their powers) a RV}
\usage{
mk_cov_rv(X, cName, across, by)

mk_cov_rv_pow(X, cName, across, by, pow = 2)
}
\arguments{
\item{X}{data.table}

\item{cName}{name of covariate}

\item{across}{RV is computed across this dimension}

\item{by}{RV is computed separately for each of these dimensions}

\item{pow}{Raise to this power. Default is 2.}
}
\value{
NULL is retured, but the data.table is modified, with appropriate columns. Original covariate column will be modified.
}
\description{
Get the mean and SD of a covariate
}
\details{
Adds columns for the covarate, it's power, and both their sd's. Columns for SD's are named by appending "_sd", column for power is named by appending the numeric value of \code{pow}.
}
\section{Functions}{
\itemize{
\item \code{mk_cov_rv_pow}: Get the mean and SD of a covariate, and the same for the covariate^pow
}}
\examples{
X <- data.table(a=rnorm(1000), b=sample(letters, 1000, replace=TRUE), key="b")
X[,K:=seq_along(a),by="b"]
mk_cov_rv_pow(X, "a", across="K", by="b", pow=2)
}

