% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_sac.R
\name{run_sac}
\alias{run_sac}
\title{Run Species Accumulation Curve Estimate of Richness}
\usage{
run_sac(regs, data_regs)
}
\arguments{
\item{regs}{Character vector of region names}

\item{data_regs}{Option list of data.tables; list should be named according to \code{regs}}
}
\description{
Calculate richness using a species accumulation curve
}

