% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_msom.R
\name{trim_msom}
\alias{trim_msom}
\title{Trim trawl data for msom}
\usage{
trim_msom(reg, gridSize = 1, plot = FALSE)
}
\arguments{
\item{reg}{name of region to be passed to \code{\link{trawlTrim}}}

\item{gridSize}{grid size to be passed to \code{\link{ll2strat}}}

\item{plot}{Logical, default FALSE; for \code{check_strat}, should the stratum tolerance be plotted?}
}
\value{
A data.table
}
\description{
Performs the \code{trawlTrim} function the clean region data, then trims down to species (not genera) that have been observed at least 10 times. Does aggregating within a haul/ spp. Adds an abundance column, which is actually just 0 if wtcpue == 0, and 1 if wtcpue > 0.
}

